+++ src/libraries/System.Private.CoreLib/src/System/IO/Enumeration/FileSystemName.cs
@@ -75,7 +75,7 @@ namespace System.IO.Enumeration
         /// <param name="name">The name to check against the expression.</param>
         /// <param name="ignoreCase"><see langword="true" /> to ignore case (default), <see langword="false" /> if the match should be case-sensitive.</param>
         /// <returns><see langword="true" /> if the given expression matches the given name; otherwise, <see langword="false" />.</returns>
-        /// <remarks>The syntax of the <paramref name="expression" /> parameter is based on the syntax used by <see cref="System.IO.FileSystemWatcher" />, which is based on [RtlIsNameInExpression](/windows/win32/devnotes/rtlisnameinexpression), which defines the rules for matching DOS wildcards (`'*'`, `'?'`, `'&lt;'`, `'&gt;'`, `'"'`).
+        /// <remarks>The syntax of the <paramref name="expression" /> parameter is based on the syntax used by FileSystemWatcher, which is based on [RtlIsNameInExpression](/windows/win32/devnotes/rtlisnameinexpression), which defines the rules for matching DOS wildcards (`'*'`, `'?'`, `'&lt;'`, `'&gt;'`, `'"'`).
         /// Matching will not correspond to Win32 behavior unless you transform the expression using <see cref="System.IO.Enumeration.FileSystemName.TranslateWin32Expression(string)" />.</remarks>
         public static bool MatchesWin32Expression(ReadOnlySpan<char> expression, ReadOnlySpan<char> name, bool ignoreCase = true)
         {
